Fix function exporting
