+++ functions/taskrouterListeners/transfersListener.private.ts
@@ -29,7 +29,7 @@ import {
   TASK_QUEUE_ENTERED,
 } from '@tech-matters/serverless-helpers/taskrouter';
 import type { TransferMeta, ChatTransferTaskAttributes } from '../transfer/helpers.private';
-import { TransitionAgentParticipants } from '../interaction/transitionAgentParticipants';
+import { InteractionChannelParticipants } from '../interaction/interactionChannelParticipants.private';
 
 export const eventTypes: EventType[] = [
   RESERVATION_ACCEPTED,
@@ -193,9 +193,9 @@ export const handleEvent = async (context: Context<EnvVars>, event: EventFields)
        * If conversation, remove original participant from conversation.
        */
 
-      const { path } = Runtime.getFunctions()['interaction/transitionAgentParticipants'];
+      const { path } = Runtime.getFunctions()['interaction/interactionChannelParticipants'];
       // eslint-disable-next-line prefer-destructuring,global-require,import/no-dynamic-require
-      const transitionAgentParticipants: TransitionAgentParticipants = require(path);
+      const { transitionAgentParticipants }: InteractionChannelParticipants = require(path);
       await transitionAgentParticipants(
         context.getTwilioClient(),
         context.TWILIO_WORKSPACE_SID,
@@ -231,9 +231,9 @@ export const handleEvent = async (context: Context<EnvVars>, event: EventFields)
        * If conversation, remove original participant from conversation.
        */
       try {
-        const { path } = Runtime.getFunctions()['interaction/transitionAgentParticipants'];
+        const { path } = Runtime.getFunctions()['interaction/interactionChannelParticipants'];
         // eslint-disable-next-line prefer-destructuring,global-require,import/no-dynamic-require
-        const transitionAgentParticipants: TransitionAgentParticipants = require(path);
+        const { transitionAgentParticipants }: InteractionChannelParticipants = require(path);
         await transitionAgentParticipants(
           context.getTwilioClient(),
           context.TWILIO_WORKSPACE_SID,
